fix issues with size parsing and conversion from uint64 to int64

+++ internal/journal/database/badger.go
@@ -40,7 +40,7 @@ func NewBadgerdb(cfg dbconfig.JournalConfig) (StorageDB, error) {
 		if err != nil {
 			return nil, err
 		}
-		opts = opts.WithValueLogFileSize(int64(size))
+		opts = opts.WithValueLogFileSize(size)
 	}
 
 	db, err := badger.Open(opts)
